Use couch db design doc view to fetch users.
