+++ src/models/User.ts
@@ -7,20 +7,26 @@ import { Device } from './Device'
 const CONFIG = require('../../serverConfig.json')
 
 const nanoDb = Nano(CONFIG.dbFullpath)
-const dbUserSettings = nanoDb.db.use('db_user_settings')
+const dbUserSettings = nanoDb.db.use<ReturnType<typeof IUser>>('db_user_settings')
 
 const IUserDevices = asMap(asBoolean)
+const IUserCurrencyCodes = asMap(asObject({
+  '1': asBoolean,
+  '24': asBoolean
+}))
 const IUserNotifications = asObject({
   enabled: asOptional(asBoolean),
-  currencyCodes: asMap(asObject({
-    '1': asBoolean,
-    '24': asBoolean
-  }))
+  currencyCodes: IUserCurrencyCodes
 })
 const IUser = asObject({
   devices: IUserDevices,
   notifications: IUserNotifications
 })
+
+export interface INotificationsEnabledViewResponse {
+  devices: ReturnType<typeof IUserDevices>
+  currencyCodes: ReturnType<typeof IUserCurrencyCodes>
+}
 
 export class User extends Base implements ReturnType<typeof IUser> {
   public static table = dbUserSettings
@@ -40,6 +46,12 @@ export class User extends Base implements ReturnType<typeof IUser> {
         currencyCodes: {}
       }
     }
+  }
+
+  // Fetch data for users that have notifications enabled using CouchDB Design Document View
+  // https://notif1.edge.app:6984/_utils/#/database/db_user_settings/_design/filter/_view/notifications-enabled
+  public static notificationsEnabled() {
+    return User.table.view<INotificationsEnabledViewResponse>('filter','notifications-enabled')
   }
 
   public async attachDevice(deviceId: string) {
